IR: keep correspondingPropertySymbol in DeepCopyIrTreeWithSymbols
JVM IR needs this information to be able to map property accessor names
correctly in codegen ("<get-foo>" -> "getFoo").

This change is not entirely correct because we keep a reference to the
"old" property symbol. Transforming it via deep-copy wouldn't be correct
either though, because then we'd end up with different copies of the
same property in its getter and setter. This is likely a temporary
solution until we have proper support for properties in the IR
(IrProperty which is IrDeclarationParent).